go: pull generated send/recv into lib
Pull the generated send/recv methods for clients into the thrift lib as
they are all the same. This greatly simplifies the methods for
generating the client code.

Add a new interface TClient which generated clients now require to
execute calls. This allows users to easily customise the behaviour of
the client, such as using a connection pool or implementing custom
clients such a THeader.